Move `ReportsSyncAndroidTest` to the `org.simple.clinic.sync` package
This commit also renames the class to `ReportsSyncIntegrationTest`.
